+++ sdk/cosmos/azure_data_cosmos/src/clients/cosmos_client.rs
@@ -29,7 +29,7 @@ pub use super::cosmos_client_builder::CosmosClientBuilder;
 /// Using Entra ID authentication:
 ///
 /// ```rust,no_run
-/// use azure_data_cosmos::{CosmosClient, CosmosAccountReference, CosmosAccountEndpoint};
+/// use azure_data_cosmos::{CosmosClient, CosmosAccountReference, CosmosAccountEndpoint, RegionSelectionStrategy, regions};
 /// use std::sync::Arc;
 ///
 /// # async fn doc() -> Result<(), Box<dyn std::error::Error>> {
@@ -40,7 +40,7 @@ pub use super::cosmos_client_builder::CosmosClientBuilder;
 ///     .unwrap();
 /// let account = CosmosAccountReference::with_credential(endpoint, credential);
 /// let client = CosmosClient::builder()
-///     .build(account)
+///     .build(account, RegionSelectionStrategy::ProximityTo(regions::EAST_US))
 ///     .await?;
 /// # Ok(())
 /// # }
@@ -49,7 +49,7 @@ pub use super::cosmos_client_builder::CosmosClientBuilder;
 /// Using key authentication (requires `key_auth` feature):
 ///
 /// ```rust,no_run,ignore
-/// use azure_data_cosmos::{CosmosClient, CosmosAccountReference, CosmosAccountEndpoint};
+/// use azure_data_cosmos::{CosmosClient, CosmosAccountReference, CosmosAccountEndpoint, RegionSelectionStrategy, regions};
 /// use azure_core::credentials::Secret;
 ///
 /// # async fn doc() -> Result<(), Box<dyn std::error::Error>> {
@@ -61,7 +61,7 @@ pub use super::cosmos_client_builder::CosmosClientBuilder;
 ///     Secret::from("my_account_key"),
 /// );
 /// let client = CosmosClient::builder()
-///     .build(account)
+///     .build(account, RegionSelectionStrategy::ProximityTo(regions::EAST_US))
 ///     .await?;
 /// # Ok(())
 /// # }
@@ -80,7 +80,7 @@ impl CosmosClient {
     /// # Examples
     ///
     /// ```rust,no_run
-    /// use azure_data_cosmos::{CosmosClient, CosmosAccountReference, CosmosAccountEndpoint};
+    /// use azure_data_cosmos::{CosmosClient, CosmosAccountReference, CosmosAccountEndpoint, RegionSelectionStrategy, regions};
     ///
     /// # async fn doc() -> Result<(), Box<dyn std::error::Error>> {
     /// let credential: std::sync::Arc<dyn azure_core::credentials::TokenCredential> =
@@ -90,7 +90,7 @@ impl CosmosClient {
     ///     .unwrap();
     /// let account = CosmosAccountReference::with_credential(endpoint, credential);
     /// let client = CosmosClient::builder()
-    ///     .build(account)
+    ///     .build(account, RegionSelectionStrategy::ProximityTo(regions::EAST_US))
     ///     .await?;
     /// # Ok(())
     /// # }